* Makefile.in (OBJS): Drop transport_sockets.o. 	* sysv_shm.cc (shmget_allocate_segment): Remove spurious cast. 	* transport.cc: Don't include transport_sockets.h. 	(create_server_transport): Always create transport_layer_pipes. 	* transport_sockets.cc: Remove. 	* transport_sockets.h: Remove.

+++ winsup/cygserver/sysv_shm.cc
@@ -729,7 +729,7 @@ shmget_allocate_segment(struct thread *td, struct shmget_args *uap, int mode)
 	if (shm_nused >= shminfo.shmmni) /* Any shmids left? */
 		return (ENOSPC);
 	size = round_page(uap->size);
-	if (shm_committed + btoc(size) > (unsigned long) shminfo.shmall)
+	if (shm_committed + btoc(size) > shminfo.shmall)
 		return (ENOMEM);
 	if (shm_last_free < 0) {
 		shmrealloc();	/* Maybe expand the shmsegs[] array. */
